aaaaaa

+++ src/edu/ecnu/dll/dataset/other/DataTest.java
@@ -10,10 +10,10 @@ import java.util.List;
 public class DataTest {
 
 
-    public static int[] getNegativeAndTotalNumberCount(String twoDimensionalDoubleDataPath) {
+    public static long[] getNegativeAndTotalNumberCount(String twoDimensionalDoubleDataPath) {
         List<Double[]>[] data = TwoDimensionDoubleRead.readDouble(twoDimensionalDoubleDataPath, 1);
-        int totalSize = 0;
-        int negativeSize = 0;
+        long totalSize = 0;
+        long negativeSize = 0;
         for (int i = 0; i < data.length; i++) {
             for (Double[] doubles : data[i]) {
                 for (int j = 0; j < doubles.length; j++) {
@@ -24,16 +24,16 @@ public class DataTest {
                 }
             }
         }
-        return new int[]{negativeSize, totalSize};
+        return new long[]{negativeSize, totalSize};
     }
 
-    public static int[] getNegativeAndTotalNumberCountInDir(String parentPath, String matchingName) {
+    public static long[] getNegativeAndTotalNumberCountInDir(String parentPath, String matchingName) {
         FilenameFilter filter = new DataFileNameFilter(matchingName);
         File parentFile = new File(parentPath);
         File[] files = parentFile.listFiles(filter);
-        int totalNumberSize = 0;
-        int totalNegativeSize = 0;
-        int[] partResult;
+        long totalNumberSize = 0;
+        long totalNegativeSize = 0;
+        long[] partResult;
         String filePath;
         for (int i = 0; i < files.length; i++) {
             filePath = files[i].getAbsolutePath();
@@ -41,7 +41,7 @@ public class DataTest {
             totalNegativeSize += partResult[0];
             totalNumberSize += partResult[1];
         }
-        return new int[]{totalNegativeSize, totalNumberSize};
+        return new long[]{totalNegativeSize, totalNumberSize};
     }
 
 
@@ -50,7 +50,7 @@ public class DataTest {
 //        String path = "F:\\test";
         String path = args[0];
         String matchingName = "noise_distance.txt";
-        int[] result = getNegativeAndTotalNumberCountInDir(path, matchingName);
+        long[] result = getNegativeAndTotalNumberCountInDir(path, matchingName);
         System.out.println(result[0]);
         System.out.println(result[1]);
         System.out.println(result[0]*1.0/result[1]);
